+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Fix
 
+- Check for non required undefined values in `predCheckable`, not in `collCheckable`, as undefined value could eventually change, but spec would still be fixed to always true.
+
 ### Deprecate
 
 ---